Tests in task 1 now use Great Circle distance
